Lot's of Changes:
Javascript:

Changed the javascript to work a little better when understanding different types of add on, and re-setting the radio button options when unchecking an add-on

Property Module:

I changed the Model for the CCK fields to have the same sort of format, my todo is to compress the code here, so I am simplifying the add on process, so one could generate/modify their own Add-ons easily.

I also added in validation, and a way for the CCK fields to form_error in deep fields.

TODOS: Compress the code where I can, and make the error elements easier to find.

Property Add-Ons Display:

Changed the logic to ignore empty fields, because even if the value is empty, PHP reconizes the $foo variable in your template no matter what, without accounting for empty() when you hit the template.